dom/file/tests/test_file_negative_date.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1158437
-->
<head>
  <title>Test for negative date in File (Bug 1158437)</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1158437">Mozilla Bug 1158437</a>

<script class="testbody" type="text/javascript">
"use strict";

var blob = new Blob(['hello world']);
var f1 = new File([blob], 'f1.txt', { lastModified: 0 });
var f2 = new File([blob], 'f2.txt', { lastModified: -1 });
var f3 = new File([blob], 'f3.txt', { lastModified: -1000 });

is(f1.lastModified, 0, "lastModified == 0 is supported");
is(f2.lastModified, -1, "lastModified == -1 is supported");
is(f3.lastModified, -1000, "lastModified == -1000 is supported");

</script>
</body>
</html>
